requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.1.0
pytesseract==0.3.10
lxml==4.9.3
selenium==4.38.0
